+++ test-data/test-config2.yml
@@ -1,6 +1,6 @@
 terraform_binary_file: ./test-plan-reader
 terraform_plan_file_basename: plan.json
-terraform_plan_search_folder: .
+terraform_plan_search_folder: ../adm.apac.spoke.prod-terragrunt/
 
 critical_resources:
   - all
